requirements.txt
python-telegram-bot==20.3
Pillow>=11.0.0

